+++ src/main/java/com/senac/projetoarquivo/repository/ProjetoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.senac.projetoarquivo.entity.Projeto;
 
-public interface ProjetoRepository extends JpaRepository<Projeto,Long>{
+public interface ProjetoRepository extends JpaRepository<Projeto,Integer>{
     
 }
